Initial commit
Usage: csv2sc < file.csv > file.sc

This program is designed to convert CSV files into sc[1] files.

It handles numerical (float and int) types and also strings and makes no
attempts to deserialize formula or anything else that sc might support.

Strings are left-aligned and escaped (using rust's debug display, this
is a potentially brittle hack). Anything that looks like a number but
can't be deserialized as such is stored as a string.

It supports parsing files with up to 702 columns, and u32::MAX rows -
which should be enough for most uses.

[1] - sc is a CLI spreadsheet program